Add intent testing and move DrawerProvider to util folder
